Low_Milk_Pattern_Detector/requirements.txt
streamlit
pandas
plotly
openai
python-dotenv
pydomo
Jinja2